+++ src/Factory/FormFactory.php
@@ -11,7 +11,7 @@ namespace Dot\User\Factory;
 
 use Dot\User\Options\UserOptions;
 use Dot\User\Options\UserOptionsAwareInterface;
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 
 /**
  * Class UserOptionsAwareDelegator